# ZadanieEshopApi/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "Bart Eshop API",
    "description": "API for managing orders and products in Bart Eshop",
    "version": "1.0.0"
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Welcome message for debugging connection",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "string",
                  "example": "Welcome in Bart Eshop API!"
                }
              }
            }
          }
        }
      }
    },
    "/orders": {
      "post": {
        "summary": "Create a new order",
        "operationId": "createOrder",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "string", "example": "customer@example.com" },
                "products": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "example": "60d5ec49f9a3c8b77c3d25a1" },
                      "quantity": { "type": "integer", "example": 2 }
                    },
                    "required": ["id", "quantity"]
                  }
                }
              },
              "required": ["email", "products"]
            }
          }
        ],
        "responses": {
          "201": { "description": "Order created" },
          "400": { "description": "Invalid request data" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "get": {
        "summary": "Get list of orders",
        "operationId": "getOrders",
        "parameters": [
          { "name": "page", "in": "query", "type": "integer", "default": 1 },
          { "name": "limit", "in": "query", "type": "integer", "default": 20 },
          { "name": "filter_email", "in": "query", "type": "string", "default": "" },
          { "name": "filter_status", "in": "query", "type": "string", "enum": ["", "unpaid", "cancelled"] }
        ],
        "responses": {
          "200": {
            "description": "List of orders",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "page": { "type": "integer" },
                    "limit": { "type": "integer" },
                    "total_orders": { "type": "integer" },
                    "total_pages": { "type": "integer" },
                    "orders": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": { "type": "string" },
                          "email": { "type": "string" },
                          "created_at": { "type": "string" },
                          "total_price": { "type": "number" },
                          "status": { "type": "string" },
                          "products": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "_id": { "type": "string" },
                                "name": { "type": "string" },
                                "price": { "type": "number" },
                                "quantity": { "type": "integer" }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/orders/{order_id}": {
      "delete": {
        "summary": "Cancel an order",
        "operationId": "cancelOrder",
        "parameters": [
          { "name": "order_id", "in": "path", "required": true, "type": "string" }
        ],
        "responses": {
          "204": { "description": "Order cancelled successfully" },
          "400": { "description": "Invalid order ID" },
          "404": { "description": "Order not found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/products": {
      "get": {
        "summary": "Get list of products",
        "operationId": "getProducts",
        "parameters": [
          { "name": "page", "in": "query", "type": "integer", "default": 1 },
          { "name": "limit", "in": "query", "type": "integer", "default": 20 },
          { "name": "name_query", "in": "query", "type": "string" },
          { "name": "order_by", "in": "query", "type": "string", "enum": ["name", "_id", "stock", "price", "units_sold"] },
          { "name": "order", "in": "query", "type": "string", "enum": ["asc", "desc"] }
        ],
        "responses": {
          "200": {
            "description": "List of products",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "page": { "type": "integer" },
                    "limit": { "type": "integer" },
                    "total_products": { "type": "integer" },
                    "total_pages": { "type": "integer" },
                    "products": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": { "type": "string" },
                          "name": { "type": "string" },
                          "price": { "type": "number" },
                          "stock": { "type": "integer" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": { "description": "Internal Server Error" }
        }
      }
    }
  },
  "definitions": {
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "error": { "type": "string" }
      }
    }
  }
}